Separate loading new file, and IC
Separate in for-loop time the loading of the next data file, and the instrument correction steps.
The result of this change is that the tr_trim data never gets too close to the edges of the st_IC, when they would be affected by the tapering of the stream that goes with the instrument correction step.
This commit still includes some additional print statements useful for debugging and monitoring flow through the script.