Make the IRIS modified version the official case
Now reassigned local version of med_spec_loop_v2 to _old, and made the IRIS version the official version.  I believe that some of the handling of the trimming and loading within the main loop that drives the action has been improved during February debugging that was part of the exploration for the Greenland_Gulley project.  When I address the pull request, to reincorpate this into the master version, I'll have to change some of the header info back to the old version, but hopefully the loop will remain improved.